style: fix lint issues (remove unuseds, rename unused params, clean sweep CLI)
