+++ cmd/parable_bloom/cmd_param_sweep.go
@@ -1,7 +1,6 @@
 package parable_bloom
 
 import (
-	"flag"
 	"fmt"
 	"time"
 )
@@ -10,20 +9,6 @@ import (
 type cfgProf struct {
 	cfg  GeneratorConfig
 	prof VarietyProfile
-}
-
-// Simple parameter sweep CLI. Run with `go run cmd/parable_bloom/cmd_param_sweep.go -difficulty Seedling -iters 5`.
-func main() {
-	difficulty := flag.String("difficulty", "Seedling", "difficulty tier to sweep")
-	iters := flag.Int("iters", 5, "number of random seeds per parameter set")
-	flag.Parse()
-
-	bestScore, bestCfg, bestProf, bestTime := SweepParams(*difficulty, *iters)
-	fmt.Printf("Sweep result for %s:\n", *difficulty)
-	fmt.Printf("  Best score: %.2f\n", bestScore)
-	fmt.Printf("  Best config: %+v\n", bestCfg)
-	fmt.Printf("  Best profile (lengthMix): %+v\n", bestProf.LengthMix)
-	fmt.Printf("  Avg elapsed ms: %d\n", bestTime)
 }
 
 // SweepParams runs the parameter sweep and returns the best results.
@@ -50,10 +35,6 @@ func SweepParams(difficulty string, iters int) (float64, GeneratorConfig, Variet
 	}
 
 	// Build candidate configurations and evaluate them
-	type cfgProf struct{
-		cfg GeneratorConfig
-		prof VarietyProfile
-	}
 	candidates := buildCandidates(seedRetries, repairRadius, repairRetries, lengthPresets)
 	for _, item := range candidates {
 		avgScore, avgTime := evaluateConfig(difficulty, spec, item.prof, item.cfg, iters)
